chore: move utils under pkg umbrellia
